requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.9.3
    # via cohere
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.2.0
    # via aiohttp
backoff==2.2.1
    # via cohere
beautifulsoup4==4.12.3
    # via bs4
blessed==1.20.0
    # via inquirer
bs4==0.0.2
    # via -r requirements.in
certifi==2024.2.2
    # via requests
charset-normalizer==3.3.2
    # via requests
cohere==4.47
    # via -r requirements.in
colorama==0.4.6
    # via
    #   halo
    #   log-symbols
ebooklib==0.18
    # via -r requirements.in
editor==1.6.6
    # via inquirer
fastavro==1.9.4
    # via cohere
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
halo==0.0.31
    # via -r requirements.in
idna==3.6
    # via
    #   requests
    #   yarl
importlib-metadata==6.11.0
    # via cohere
inquirer==3.2.3
    # via -r requirements.in
log-symbols==0.0.14
    # via halo
lxml==5.1.0
    # via ebooklib
multidict==6.0.5
    # via
    #   aiohttp
    #   yarl
numpy==1.26.4
    # via -r requirements.in
python-dotenv==1.0.1
    # via -r requirements.in
readchar==4.0.5
    # via inquirer
requests==2.31.0
    # via cohere
runs==1.2.2
    # via editor
six==1.16.0
    # via
    #   blessed
    #   ebooklib
    #   halo
soupsieve==2.5
    # via beautifulsoup4
spinners==0.0.24
    # via halo
termcolor==2.4.0
    # via halo
tqdm==4.66.2
    # via -r requirements.in
urllib3==2.2.1
    # via
    #   cohere
    #   requests
wcwidth==0.2.13
    # via blessed
xmod==1.8.1
    # via
    #   editor
    #   runs
yarl==1.9.4
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# setuptools
